Deep dive: compile/verify Java 43-52, expand DS 31-35, Algo 36-40

+++ labs/java/43-class-loading-bytecode/src/main/java/com/javaacademy/lab43/classloading/BytecodeAnalyzer.java
@@ -46,7 +46,7 @@ public class BytecodeAnalyzer {
 
         int ifCount = dis.readUnsignedShort();
         System.out.println("Interface Count: " + ifCount);
-        dis.skipBytes(ifCount * 2L);
+        dis.skipBytes((int) (ifCount * 2L));
 
         int fieldCount = dis.readUnsignedShort();
         System.out.println("Field Count: " + fieldCount);
